+++ nowcast_lstm/LSTM.py
@@ -5,6 +5,7 @@ import datetime
 
 import nowcast_lstm.data_setup
 import nowcast_lstm.modelling
+import nowcast_lstm.interval_prediction as ip
 
 
 class LSTM:
@@ -20,7 +21,7 @@ class LSTM:
     `model.gen_ragged_X(pub_lags, lag)` to generate a data vintage of X model input, useful for evaluation, how would this model have performed historically with missing data.
 
 
-        parameters:
+    parameters:
         :data: pandas DataFrame: n x m+1 dataframe
         :target_variable: str: name of the target var
         :n_timesteps: how many historical periods to consider when training the model. For example if the original data is monthly, n_steps=12 would consider data for the last year.
@@ -112,7 +113,7 @@ class LSTM:
 
         :num_workers: int: number of workers for multi-process data loading
         :shuffle: boolean: whether to shuffle data at every epoch
-                :quiet: boolean: whether or not to print the losses in the epoch loop
+        :quiet: boolean: whether or not to print the losses in the epoch loop
         """
         for i in range(self.n_models):
             if quiet == False:
@@ -152,11 +153,11 @@ class LSTM:
         """Make predictions on a dataframe with same columns and order as the model was trained on, including the target variable.
 
         parameters:
-                :data: pandas DataFrame: data to predict fitted model on
+            :data: pandas DataFrame: data to predict fitted model on
             :only_actuals_obs: boolean: whether or not to predict observations without a target actual
 
         output:
-                :return: pandas DataFrame of dates and predictions
+            :return: pandas DataFrame of dates and predictions
         """
         dataset = self.data_setup.gen_dataset(
             data,
@@ -204,7 +205,7 @@ class LSTM:
             :end_date: str in "YYYY-MM-DD" format: end date of generating ragged preds
 
         output:
-                :return: numpy array equivalent in shape to X input, but with trailing edges set to missing/0, ys, and dates
+            :return: numpy array equivalent in shape to X input, but with trailing edges set to missing/0, ys, and dates
         """
         if data is None:
             data = self.data
@@ -266,14 +267,14 @@ class LSTM:
         """Get predictions on artificial vintages
 
         parameters:
-                :pub_lags: list[int]: list of periods back each input variable is set to missing. I.e. publication lag of the variable.
+            :pub_lags: list[int]: list of periods back each input variable is set to missing. I.e. publication lag of the variable.
             :lag: int: simulated periods back. E.g. -2 = simulating data as it would have been 2 months before target period, 1 = 1 month after, etc.
             :data: pandas DataFrame: dataframe to generate the ragged datasets on, if none will calculate on training data
             :start_date: str in "YYYY-MM-DD" format: start date of generating ragged preds. To save calculation time, i.e. just calculating after testing date instead of all dates
             :end_date: str in "YYYY-MM-DD" format: end date of generating ragged preds
 
         output:
-                :return: pandas DataFrame of actuals, predictions, and dates
+            :return: pandas DataFrame of actuals, predictions, and dates
         """
         X, y, dates = self.gen_ragged_X(pub_lags, lag, data, start_date, end_date)
         # predictions on every model
@@ -323,3 +324,233 @@ class LSTM:
             :scaled_contribution: contribution of feature to the model, scaled to 1 = most important feature
         """
         return self.modelling.feature_contribution(self)
+
+    def interval_predict(
+        self,
+        data,
+        interval=0.95,
+        only_actuals_obs=False,
+        start_date=None,
+        end_date=None,
+    ):
+        """Get predictions plus uncertainty intervals on a new dataset
+
+        parameters:
+            :data: pandas DataFrame: data to predict fitted model on
+            :interval: float: float between 0 and 1, uncertainty interval. A closer number to one gives a higher uncertainty interval. E.g., 0.95 (95%) will give larger bands than 0.5 (50%)
+            :only_actuals_obs: boolean: whether or not to predict observations without a target actual
+            :start_date: str in "YYYY-MM-DD" format: start date of generating interval predictions. To save calculation time, i.e. just calculating after testing date instead of all dates
+            :end_date: str in "YYYY-MM-DD" format: end date of generating interval predictions
+
+        output:
+            :return: pandas DataFrame of actuals, point predictions, lower and upper uncertainty intervals.
+        """
+        # filter for desired dates
+        if start_date is None:
+            data_start_index = 0
+        else:
+            data_start_index = (
+                data.index[data.date >= start_date].tolist()[0] - self.n_timesteps + 1
+            )
+
+        if end_date is None:
+            end_date = str(np.max(data.date))[:10]
+
+        cut_data = data.iloc[data_start_index:, :].reset_index(drop=True)
+        cut_data = cut_data.loc[cut_data.date <= end_date, :].reset_index(drop=True)
+
+        dataset = self.data_setup.gen_dataset(
+            cut_data,
+            self.target_variable,
+            self.fill_na_func,
+            self.fill_ragged_edges_func,
+            fill_na_other_df=self.data,  # use training data to calculate fill na means, etc.
+            arma_full_df=data,  # use this data to fit ARMA model on
+        )
+        na_filled_dataset = dataset["na_filled_dataset"]
+        date_series = dataset["date_series"]
+        model_input = self.data_setup.gen_model_input(
+            na_filled_dataset, self.n_timesteps, drop_missing_ys=False
+        )
+        X = model_input[0]
+        y = model_input[1]
+
+        # predictions on every model
+        preds = []
+        for i in range(self.n_models):
+            preds.append(self.modelling.predict(X, self.mv_lstm[i]))
+
+        point_preds = list(np.mean(preds, axis=0))
+
+        prediction_df = pd.DataFrame(
+            {
+                "date": date_series[
+                    (len(date_series) - len(point_preds)) :
+                ].values.flatten(),  # may lose some observations at the beginning depending on n_timeperiods, account for that
+                "actuals": y,
+                "point_predictions": point_preds,
+                "lower_interval": np.nan,
+                "upper_interval": np.nan,
+            }
+        )
+
+        # data availabilities
+        availabilities = [
+            ip.calc_perc_available(self, data, target_period)
+            for target_period in prediction_df.date
+        ]
+
+        # standard deviations of the chosen interval (from a normal distribution)
+        sds = ip.calc_sds(interval)
+
+        # standard deviation of target variable
+        target_sd = np.std(data[self.target_variable])
+
+        # predictions
+        interval_preds = [
+            ip.single_interval_predict(
+                sample=np.array(preds)[:, i],
+                sds=sds,
+                target_sd=target_sd,
+                availability=availabilities[i],
+                interval=interval,
+            )
+            for i in range(len(availabilities))
+        ]
+
+        final_df = pd.DataFrame(
+            {
+                "date": prediction_df.date,
+                "actuals": prediction_df.actuals,
+                "predictions": [x[0] for x in interval_preds],
+                "lower_interval": [x[1] for x in interval_preds],
+                "upper_interval": [x[2] for x in interval_preds],
+            }
+        )
+
+        if only_actuals_obs:
+            final_df = final_df.loc[~pd.isna(final_df.actuals), :].reset_index(
+                drop=True
+            )
+
+        return final_df
+
+    def ragged_interval_predict(
+        self,
+        pub_lags,
+        lag,
+        data,
+        interval=0.95,
+        start_date=None,
+        end_date=None,
+    ):
+        """Get predictions plus uncertainty intervals on artificial vintages
+
+        parameters:
+            :pub_lags: list[int]: list of periods back each input variable is set to missing. I.e. publication lag of the variable.
+            :lag: int: simulated periods back. E.g. -2 = simulating data as it would have been 2 months before target period, 1 = 1 month after, etc.
+            :data: pandas DataFrame: data to predict fitted model on
+            :interval: float: float between 0 and 1, uncertainty interval. A closer number to one gives a higher uncertainty interval. E.g., 0.95 (95%) will give larger bands than 0.5 (50%)
+            :start_date: str in "YYYY-MM-DD" format: start date of generating interval predictions. To save calculation time, i.e. just calculating after testing date instead of all dates
+            :end_date: str in "YYYY-MM-DD" format: end date of generating interval predictions
+
+        output:
+            :return: pandas DataFrame of actuals, point predictions, lower and upper uncertainty intervals.
+        """
+        # filter for desired dates
+        if start_date is None:
+            data_start_index = 0
+            start_date = str(np.min(data.date))[:10]
+        else:
+            data_start_index = (
+                data.index[data.date >= start_date].tolist()[0] - self.n_timesteps + 1
+            )
+
+        if end_date is None:
+            end_date = str(np.max(data.date))[:10]
+
+        cut_data = data.iloc[data_start_index:, :].reset_index(drop=True)
+        cut_data = cut_data.loc[cut_data.date <= end_date, :].reset_index(drop=True)
+
+        X, y, date_series = self.gen_ragged_X(
+            pub_lags=pub_lags,
+            lag=lag,
+            data=cut_data,
+            start_date=start_date,
+            end_date=end_date,
+        )
+        date_series = pd.Series(date_series)
+
+        # predictions on every model
+        preds = []
+        for i in range(self.n_models):
+            preds.append(self.modelling.predict(X, self.mv_lstm[i]))
+
+        point_preds = list(np.mean(preds, axis=0))
+
+        prediction_df = pd.DataFrame(
+            {
+                "date": date_series[
+                    (len(date_series) - len(point_preds)) :
+                ].values.flatten(),  # may lose some observations at the beginning depending on n_timeperiods, account for that
+                "actuals": y,
+                "point_predictions": point_preds,
+                "lower_interval": np.nan,
+                "upper_interval": np.nan,
+            }
+        )
+
+        # data availabilities
+        # function to lag data
+        def lag_data(target_variable, pub_lags, data, last_date, lag):
+            import numpy as np
+
+            final = data.loc[data.date <= last_date, :].reset_index(drop=True)
+            tmp = final.drop(["date", target_variable], axis=1)
+            for i in range(len(tmp.columns)):
+                tmp_lag = pub_lags[i]
+                # go back as far as needed for the pub_lag of the variable, then + the lag (so -2 for 2 months back), also -1 because 0 lag means in month, last month data available, not current month in
+                final.loc[(len(final) - tmp_lag + lag - 1) :, tmp.columns[i]] = np.nan
+            return final
+
+        availability_dfs = []
+        for i in range(len(prediction_df.date)):
+            availability_df = lag_data(
+                self.target_variable, pub_lags, data, prediction_df.date[i], lag
+            )
+            availability_dfs.append(availability_df)
+
+        availabilities = [
+            ip.calc_perc_available(self, availability_dfs[i], prediction_df.date[i])
+            for i in range(len(prediction_df.date))
+        ]
+
+        # standard deviations of the chosen interval (from a normal distribution)
+        sds = ip.calc_sds(interval)
+
+        # standard deviation of target variable
+        target_sd = np.std(data[self.target_variable])
+
+        # predictions
+        interval_preds = [
+            ip.single_interval_predict(
+                sample=np.array(preds)[:, i],
+                sds=sds,
+                target_sd=target_sd,
+                availability=availabilities[i],
+                interval=interval,
+            )
+            for i in range(len(availabilities))
+        ]
+
+        final_df = pd.DataFrame(
+            {
+                "date": prediction_df.date,
+                "actuals": prediction_df.actuals,
+                "predictions": [x[0] for x in interval_preds],
+                "lower_interval": [x[1] for x in interval_preds],
+                "upper_interval": [x[2] for x in interval_preds],
+            }
+        )
+        
+        return final_df